authorize permissions

+++ app/Http/Controllers/Auth/RoleController.php
@@ -8,13 +8,17 @@ use App\Services\Auth\RoleService;
 use App\Services\Auth\UserService;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Support\Facades\Redirect;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Spatie\Permission\Models\Permission;
+use Spatie\Permission\Models\Role;
 use Inertia\Inertia;
 use Inertia\Response;
 use App\Models\User;
 
 class RoleController extends Controller
 {
+    use AuthorizesRequests;
+
     private $roleService;
     private $userService;
 
@@ -26,6 +30,7 @@ class RoleController extends Controller
 
     public function index(Request $request): Response
     {
+        $this->authorize('admin roles:read', Role::class);
         $users = User::select('id', 'name')->get();
         $roles = $this->roleService->getRoles();
         $permissions = Permission::select('id', 'name')->get();
@@ -40,12 +45,14 @@ class RoleController extends Controller
 
     public function store(Request $request): RedirectResponse
     {
+        $this->authorize('admin roles:create', Role::class);
         $this->roleService->createRole($request);
         return Redirect::route('role.index');
     }
 
     public function update(Request $request): RedirectResponse
     {
+        $this->authorize('admin roles:write', Role::class);
         $this->roleService->updateRole($request);
         return Redirect::route('role.index');
     }
@@ -55,6 +62,7 @@ class RoleController extends Controller
      */
     public function destroy(Request $request): RedirectResponse
     {
+        $this->authorize('admin roles:write', Role::class);
         $this->roleService->deleteRole($request);
         return Redirect::route('role.index');
     }
